links_shortener/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.0
psycopg2-binary==2.9.8
python-dotenv==1.0.1